0903 오전
컴포넌트 모듈화 수정, 검색기능 노출되게 수정
