+++ Source/SemiProject/src/main/reactproject/src/components/Footer.js
@@ -1,0 +1,9 @@
+function Footer() {
+  return (
+    <footer>
+      <p>© 2024 게시판 웹사이트. All rights reserved.</p>
+    </footer>
+  );
+}
+
+export default Footer;
