--- Source/SemiProject/src/main/reactproject/src/pages/Footer.js
@@ -1,10 +1,0 @@
-function Footer(){
-  return (
-    <footer>
-      <hr/>
-      푸터 입니다.
-    </footer>
-  );
-};
-
-export default Footer;
