--- Source/SemiProject/src/main/reactproject/src/pages/Header.js
@@ -1,10 +1,0 @@
-function Header(){
-  return (
-    <header>
-      헤더입니다.
-      <hr/>
-    </header>
-  );
-};
-
-export default Header;
